src/main/java/am/runningtracker/service/UserService.java
package am.runningtracker.service;

import am.runningtracker.model.User;
import am.runningtracker.repo.UserRepository;
import am.runningtracker.exception.UserNotFoundException;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UserService {
    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User addUser(User user) {
        return userRepository.save(user);
    }

    public User getUser(Long id) {
        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User with id " + id + " not found"));
    }

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public void deleteUser(Long id) {
        userRepository.deleteById(id);
    }

    public User updateUser(Long id, User userUpdate) {
        User existingUser = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User with id " + id + " not found"));

        if (userUpdate.getFirstName() != null) {
            existingUser.setFirstName(userUpdate.getFirstName());
        }
        if (userUpdate.getLastName() != null) {
            existingUser.setLastName(userUpdate.getLastName());
        }
        if (userUpdate.getBirthDate() != null) {
            existingUser.setBirthDate(userUpdate.getBirthDate());
        }
        if (userUpdate.getSex() != null) {
            existingUser.setSex(userUpdate.getSex());
        }

        return userRepository.save(existingUser);
    }
}